invoice dialog finished

+++ client/src/app/components/bills/bills.component.ts
@@ -28,7 +28,8 @@ export class BillsComponent implements OnInit {
   invoices: Invoice[] = [];
   quoteSelection = new SelectionModel<Quote>(true, []);
   invoiceSelection = new SelectionModel<Invoice>(true, []);
-  errorMessage = '';
+  errorMessageAtLeastOne = '';
+  errorMessageLockedInvoices = '';
   quote: Quote = new Quote();
   invoice: Invoice = new Invoice();
 
@@ -63,6 +64,10 @@ export class BillsComponent implements OnInit {
     this.displayedQuoteColumns = (this.mobileQuery.matches) ? ['select', 'code', 'customer', 'info'] : ['select', 'code', 'subject', 'customer', 'dateQuote', 'datePayment', 'totalPriceET', 'status', 'info'];
     this.displayedInvoiceColumns = (this.mobileQuery.matches) ? ['select', 'code', 'customer', 'info'] : ['select', 'code', 'subject', 'customer', 'dateInvoice', 'datePayment', 'totalPriceET', 'status', 'info'];
     this.getQuotes();
+    this.translate.get(['bills']).subscribe(translation => {
+      this.errorMessageAtLeastOne = translation.bills.select_at_least_one;
+      this.errorMessageLockedInvoices = translation.bills.cant_delete_locked;
+    })
   }
 
   getQuotes() {
@@ -149,17 +154,14 @@ export class BillsComponent implements OnInit {
 
   onDeleteQuoteClick() {
     if (this.quoteSelection.isEmpty()) {
-      this.translate.get(['bills']).subscribe(translation => {
-        this.errorMessage = translation.bills.select_at_least_one;
-        this.alertService.error(this.errorMessage);
-      })
+        this.alertService.error(this.errorMessageAtLeastOne);
     } else {
       this.dataService.deleteQuotes(this.quoteSelection.selected).subscribe(
         data => {
           this.quoteSelection.selected.forEach(item => {
             let index: number = this.quoteDataSource.data.findIndex(i => i === item);
             this.quoteDataSource.data.splice(index, 1);
-          })
+          });
           this.quoteDataSource._updateChangeSubscription();
           this.quoteSelection.clear();
           this.alertService.success(data.message);
@@ -172,26 +174,33 @@ export class BillsComponent implements OnInit {
   }
 
   onDeleteInvoiceClick() {
+    let cancel:boolean=false;
     if (this.invoiceSelection.isEmpty()) {
-      this.translate.get(['bills']).subscribe(translation => {
-        this.errorMessage = translation.bills.select_at_least_one;
-        this.alertService.error(this.errorMessage);
-      })
-    } else {
-      this.dataService.deleteInvoices(this.invoiceSelection.selected).subscribe(
-        data => {
-          this.invoiceSelection.selected.forEach(item => {
-            let index: number = this.invoiceDataSource.data.findIndex(i => i === item);
-            this.invoiceDataSource.data.splice(index, 1);
-          })
-          this.invoiceDataSource._updateChangeSubscription();
-          this.invoiceSelection.clear();
-          this.alertService.success(data.message);
-        },
-        error => {
-          this.alertService.error(error.error.message);
+        this.alertService.error(this.errorMessageAtLeastOne);
+    }  else {
+      this.invoiceSelection.selected.forEach(item => {
+        if (item.status !== 'draft') {
+          cancel = true;
         }
-      )
+      })
+      if (cancel) {
+        this.alertService.error(this.errorMessageLockedInvoices);
+      } else {
+        this.dataService.deleteInvoices(this.invoiceSelection.selected).subscribe(
+          data => {
+            this.invoiceSelection.selected.forEach(item => {
+              let index: number = this.invoiceDataSource.data.findIndex(i => i === item);
+              this.invoiceDataSource.data.splice(index, 1);
+            })
+            this.invoiceDataSource._updateChangeSubscription();
+            this.invoiceSelection.clear();
+            this.alertService.success(data.message);
+          },
+          error => {
+            this.alertService.error(error.error.message);
+          }
+        )
+      }
     }
   }
 
@@ -223,6 +232,9 @@ export class BillsComponent implements OnInit {
         } else if (result.createMode) {
           this.quoteDataSource.data.push(result.data);
           this.quoteDataSource._updateChangeSubscription();
+        } else if(result.generateInvoiceMode){
+          this.invoiceDataSource.data.push(result.data);
+          this.invoiceDataSource._updateChangeSubscription();
         }
         this.alertService.success(result.message);
       }
@@ -238,15 +250,15 @@ export class BillsComponent implements OnInit {
     let dialogRef = this.dialog.open(InvoiceDialogComponent, config);
     dialogRef.afterClosed().subscribe(result => {
       if (result) {
-        this.alertService.success(result.message);
         if (result.editMode) {
-          let index: number = this.invoiceDataSource.data.findIndex(i => i.code === result.data.code);
-          this.invoiceDataSource.data.splice(index, 1, result.data);
-          this.invoiceDataSource._updateChangeSubscription();
-        } else {
+          this.updateInvoiceDataTable(result.data);
+        } else if (result.sendMode) {
+          this.sendInvoice(result.data);
+        } else if (result.createMode) {
           this.invoiceDataSource.data.push(result.data);
           this.invoiceDataSource._updateChangeSubscription();
         }
+        this.alertService.success(result.message);
       }
     });
   }
